Updated the multiengine and task interface documentation.
The multiengine interface documentation is in very good condition.

The task interface documentation is in OK condition.  The main problem
is that the task documentation had fallen very out of date.  Rather
than trying to update all of it, I moved the old task documentation
to a file (docs/source/parallel/parallel_task_old.txt) and started
a new shorter task documentation that is fully up to date.

This should get user's started with the new stuff.  But, we do need
to write more complete task interface docs.