Fix rpmlint: non-executable-script
"""
This text file contains a shebang or is located in a path dedicated for
executables, but lacks the executable bits and cannot thus be executed.  If
the file is meant to be an executable script, add the executable bits,
otherwise remove the shebang or move the file elsewhere.
"""

Mostly deleting the shebang, but some files contain a __main__ function,
so make them executable.

This is the last commit of this series and:
Closes gh-647.