ipcontroller cleans up connection files unless reuse=True
Connection files are not valid across sessions if reuse is False, but were previously preserved, which could cause engines to try to connect to the wrong ports if the Controller took too long to start and/or write new connection files.

Also shuffles signal-handling around a bit, so that a clean exit occurs, rather than calling sys.exit directly in the signal handler.