Fix use_dill() call
The use_dill() function is in ipyparallel.serialize.canning. Because
the ipyparallel __init__.py has a line "from serialize import *", we
can (and have to) access canning via ipyparallel.canning. This commit
fixes calling use_dill(). Currently, calling use_dill() on a direct
view is broken as it tries calling ipyparallel.serialize.use_dill(),
which is not defined. This commit fixes that, and also updates the
documentation.