Add configuration file
This builds on https://github.com/dask/dask/pull/3432 to centralize
configuration within Dask.

This adds a new jobqueue.yaml config file that allows users to specify
configuration in a file and have that be used across all of their use of the
dask-jobqueue project.  This is also something that might allow IT to help
standardize configuration options for a given cluster.