Replace threads with cores and add memory parameter (#86)
Previously people would provide the threads, processes, and memory-per-process.
Now we ask them to provide the total number of cores, total amount of memory
and number of processes by which to cut this up.  This set of parameters is somewhat
easier for new users to handle, because they already know the specifications
of their nodes.  

This removes the old threads parameter and redefine memory to memory-of-job.

This also adds documentation