Use new serialization code for PyPy3.10+ 7.3.19+
Enable the new `inspect.signature(types.CodeType)` code path for
PyPy3.10 7.3.19 and newer, notably including PyPy3.11 that does not work
with the fallback code anymore.  Add a workaround to skip the `magic`
parameter that does not seem to be exposed (or needed).

With these changes, the serialization tests pass on PyPy3.11 7.3.19,
as well as most of the test suite.

Fixes #933

Signed-off-by: Michał Górny <mgorny@gentoo.org>